fix psalm issues
