+++ src/Faker/FakerEnumProvider.php
@@ -11,6 +11,7 @@ class FakerEnumProvider extends Base
 {
     public static function register(): void
     {
+        /** @var FakerGenerator $faker */
         $faker = app(FakerGenerator::class);
 
         $providers = array_map('get_class', $faker->getProviders());
